Flatten util exports
