+++ src/Layer/brushes.js
@@ -3,7 +3,7 @@
 import {map, flatten, isNumber} from 'lodash'
 
 import {BACKGROUND_OFFSET} from '../chartCore/defaults'
-import {getPath2D} from '../utils/path2DUtils'
+import getPath2D from '../utils/getPath2D'
 
 import {
   bottomHorizontalLine,
