+++ src/utils/rect/getMinX.js
@@ -1,0 +1,5 @@
+// Copyright 2018 Kensho Technologies, LLC.
+
+export default function getMinX(rectInput) {
+  return rectInput.x
+}
