+++ src/utils/rect/getMinY.js
@@ -1,0 +1,5 @@
+// Copyright 2018 Kensho Technologies, LLC.
+
+export default function getMinY(rectInput) {
+  return rectInput.y
+}
