+++ src/utils/rect/rectBase.js
@@ -1,0 +1,3 @@
+// Copyright 2018 Kensho Technologies, LLC.
+
+export default {x: 0, y: 0, width: 0, height: 0}
