+++ test/CanvasRender/clearAndClip.js
@@ -4,7 +4,7 @@ import assert from 'assert'
 
 import {it as test} from 'mocha'
 
-import {ctxMock} from '../../src/utils/canvasUtils'
+import ctxMock from '../../src/utils/ctxMock'
 import clearAndClip from '../../src/CanvasRender/clearAndClip'
 
 const plotRect = {x: 0, y: 0, width: 500, height: 500}
